Refactoring logic for popping pages based on hint Refactored hierarchy print helper to separate class
